add central error handling as well as refactor validation error handling
