+++ routes/index.js
@@ -1,9 +1,12 @@
 const express = require('express');
 const articlesRouter = require('./articles');
 const usersRouter = require('./users');
-const handleValidationErrors = require('../middleware/error/validation');
+
 const auth = require('../controllers/auth');
 const { handleCreateUser } = require('../controllers/users');
+
+const validationErrorHandler = require('../middleware/error/validation-error-handler');
+const centralErrorHandler = require('../middleware/error/central-error-handler');
 
 const validateSignUpRequest = require('../middleware/validation/user');
 const validateSignInRequest = require('../middleware/validation/auth');
@@ -18,6 +21,7 @@ router.post('/signup', validateSignUpRequest, handleCreateUser);
 router.use('/articles', articlesRouter);
 router.use('/users', usersRouter);
 
-router.use(handleValidationErrors);
+router.use(validationErrorHandler);
+router.use(centralErrorHandler);
 
 module.exports = router;
